requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
boto3==1.34.0
paramiko==3.3.1
pyyaml==6.0.1
pydantic==2.5.0
python-multipart==0.0.6
aiofiles==23.2.1